run script checks if OS X before installing homebrew

+++ run.sh
@@ -3,12 +3,11 @@ exists () {
     type "$1" >/dev/null 2>/dev/null
 }
 
-if ! exists brew; then
-    # TODO should make sure system is posix
+if ! exists brew && [ "$(uname)" == 'Darwin' ]; then
     echo "installing homebrew"
     ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 fi
-if ! exists spark-submit; then
+if ! exists spark-submit && [ "$(uname)" == 'Darwin' ]; then
     echo "installing spark via homebrew"
     brew install apache-spark
 fi
